+++ src/time.rs
@@ -3,7 +3,7 @@ use core::convert::TryFrom;
 use core::fmt::Debug;
 
 #[cfg(feature = "chrono")]
-use chrono::{self, Datelike, Local, TimeZone, Timelike};
+use chrono::{self, Datelike, Timelike};
 
 const MIN_YEAR: u16 = 1980;
 const MAX_YEAR: u16 = 2107;
@@ -131,27 +131,29 @@ impl DateTime {
 }
 
 #[cfg(feature = "chrono")]
-impl From<Date> for chrono::Date<Local> {
+impl From<Date> for chrono::NaiveDate {
     fn from(date: Date) -> Self {
-        Local.ymd(i32::from(date.year), u32::from(date.month), u32::from(date.day))
+        chrono::NaiveDate::from_ymd_opt(i32::from(date.year), u32::from(date.month), u32::from(date.day)).unwrap()
     }
 }
 
 #[cfg(feature = "chrono")]
-impl From<DateTime> for chrono::DateTime<Local> {
+impl From<DateTime> for chrono::NaiveDateTime {
     fn from(date_time: DateTime) -> Self {
-        chrono::Date::<Local>::from(date_time.date).and_hms_milli(
-            u32::from(date_time.time.hour),
-            u32::from(date_time.time.min),
-            u32::from(date_time.time.sec),
-            u32::from(date_time.time.millis),
-        )
+        chrono::NaiveDate::from(date_time.date)
+            .and_hms_milli_opt(
+                u32::from(date_time.time.hour),
+                u32::from(date_time.time.min),
+                u32::from(date_time.time.sec),
+                u32::from(date_time.time.millis),
+            )
+            .unwrap()
     }
 }
 
 #[cfg(feature = "chrono")]
-impl From<chrono::Date<Local>> for Date {
-    fn from(date: chrono::Date<Local>) -> Self {
+impl From<chrono::NaiveDate> for Date {
+    fn from(date: chrono::NaiveDate) -> Self {
         #[allow(clippy::cast_sign_loss)]
         let year = u16::try_from(date.year()).unwrap(); // safe unwrap unless year is below 0 or above u16::MAX
         assert!((MIN_YEAR..=MAX_YEAR).contains(&year), "year out of range");
@@ -164,8 +166,8 @@ impl From<chrono::Date<Local>> for Date {
 }
 
 #[cfg(feature = "chrono")]
-impl From<chrono::DateTime<Local>> for DateTime {
-    fn from(date_time: chrono::DateTime<Local>) -> Self {
+impl From<chrono::NaiveDateTime> for DateTime {
+    fn from(date_time: chrono::NaiveDateTime) -> Self {
         let millis_leap = date_time.nanosecond() / 1_000_000; // value in the range [0, 1999] (> 999 if leap second)
         let millis = millis_leap.min(999); // during leap second set milliseconds to 999
         let date = Date::from(date_time.date());
@@ -207,11 +209,11 @@ impl ChronoTimeProvider {
 #[cfg(feature = "chrono")]
 impl TimeProvider for ChronoTimeProvider {
     fn get_current_date(&self) -> Date {
-        Date::from(chrono::Local::now().date())
+        Date::from(chrono::Local::now().date_naive())
     }
 
     fn get_current_date_time(&self) -> DateTime {
-        DateTime::from(chrono::Local::now())
+        DateTime::from(chrono::Local::now().naive_local())
     }
 }
 
@@ -298,8 +300,10 @@ mod tests {
 
     #[test]
     fn date_time_from_chrono_leap_second() {
-        use super::TimeZone;
-        let chrono_date_time = super::Local.ymd(2016, 12, 31).and_hms_milli(23, 59, 59, 1999);
+        let chrono_date_time = chrono::NaiveDate::from_ymd_opt(2016, 12, 31)
+            .unwrap()
+            .and_hms_milli_opt(23, 59, 59, 1999)
+            .unwrap();
         let date_time = DateTime::from(chrono_date_time);
         assert_eq!(
             date_time,